Support user supplied socket paths for restricted environments
Restricted shell environments are those which sanitize the OS
environment and effectively prevent `$BEAM_NOTIFY_OPTIONS` from being
passed. As a result, the user needs to find some other way to pass the
socket path to the `beam_notify` binary. This adds the `:path` option to
starting the `BEAMNotify` GenServer and the `-p` option to `beam_notify`
so that they can hardcode a path or do whatever makes sense for their
situation.

A side effect of this change is that `$BEAM_NOTIFY_OPTIONS` now mirrors
the commandline parameters. When in doubt, users can look at it for a
hint.