Create filesystem-safe names for socket files
This calls `:erlang.phash2/1` on the names to sanitize the names. This was
chosen since it's part of the Erlang stdlib. The hash is only 27-bits, but that
seems like it should be sufficient to avoid collision since the number of
active sockets at a time is likely to be only a few at max.